BUG - Search pagination, item per page and count
This line don't have to exist. 
On our website we have categories with more than 1K ads and 10 ads per page. But with this line we can't have more than 1K ads displayed. Users can't go to the 101 page, because pagination is set with this count.

It doesn't need to have limit to the count case.

$this->dao->limit(100*$this->results_per_page);