Add filter in search to add data from plugins to items
Hi,

Just an idea but so useful for plugins. <3

If found this line on item controller :
``PHP
$item = osc_apply_filter('pre_show_item', $this->itemManager->findByPrimaryKey( Params::getParam('id') ));
```

And it's perfect to complete an item with data from plugins with array_merge.  After this I create an helper in my plugin who use osc_item_field('plugin_filed', locale).

This new line will allow users to use plugins helpers to display new informations on search.

Regards