Pattern search work with all languages simultaneously
I removed these lines to be able to search a pattern simultaneously with all languages. (the last line was also very strange and can't work with more than one language)

It's strange to search a pattern only for the current language of the user. 

When I set my language to english and I'm on a French website, I want to be able to search a pattern on french listings and not only on english listings.   

It's also very useful for the admin. 


 if(empty($this->locale_code)) {
         if(OC_ADMIN) {
            $this->locale_code[osc_current_admin_locale()] = osc_current_admin_locale();
          } else {
           $this->locale_code[osc_current_user_locale()] = osc_current_user_locale();
           }
       }
$this->dao->where(sprintf("( d.fk_c_locale_code LIKE '%s' )", implode("' d.fk_c_locale_code LIKE '", $this->locale_code)));