Remove CSRF check on item add post and edit post
Hi,

Maybe it's not the best solution, but doing a csrf on item post and edit isn't really friendly. We have so many users who are losing their Description because they are posting an item and with an other tab they are surfing on our website. And if they cross an other form with csrf check, they lose their token from the session. 

So when they come back on their description, they click on post or edit my item and the lose all their modifications. 
And sometimes they have write their description during more than 10 minutes.

Any best idea ?

Regards,